refactor: merge divergent FakeMojangAPI (common + mc1.12.2 superset)
